+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-seo-meta-box` will be documented in this file
 
+## v1.0.2 - 2020-01-26
+
+- Fix model attribution based on class name
+- Added checks on request seo_title and seo_description to allow https://github.com/gabrieliuga/laravel-nova-field-seo-metabox to function
+
 ## v1.0.1 - 2020-01-21
 
 + add auto-fill for title and description if the model defines it